SimpleWebApp/pyproject.toml
[project]
name = "simplewebapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
]
